+++ showcase/harness/src/shared/cell-model/cell-model.equivalence.test.ts
@@ -1,0 +1,50 @@
+/**
+ * Behavior-preservation gate for the `buildCellModel` relocation.
+ *
+ * The pure cell-classification fold cluster (`cell-model` / `live-status` /
+ * `staleness` / `format-ts`) moved OUT of `showcase/shell-dashboard/src/lib/`
+ * INTO this harness `shared/cell-model/` directory so the dashboard AND a new
+ * harness monitor can import ONE copy. This test proves the move changed NO
+ * behavior: it runs `buildCellModel` from the NEW harness location over the
+ * committed fixtures and asserts the serialized output is byte-identical to the
+ * pre-move baseline (`cell-model.equivalence-baseline.json`), which was
+ * generated by running the ORIGINAL shell-dashboard code (git HEAD, before the
+ * move) over those SAME fixtures.
+ */
+import { describe, it, expect } from "vitest";
+import { readFileSync } from "node:fs";
+import { fileURLToPath } from "node:url";
+import { dirname, join } from "node:path";
+import { buildCellModel } from "./cell-model";
+import {
+  FIXTURES,
+  NOW,
+  serializeModel,
+} from "./cell-model.equivalence-fixtures";
+
+const HERE = dirname(fileURLToPath(import.meta.url));
+const BASELINE: Record<string, unknown> = JSON.parse(
+  readFileSync(join(HERE, "cell-model.equivalence-baseline.json"), "utf8"),
+);
+
+describe("buildCellModel relocation — behavior equivalence", () => {
+  it("produces output byte-identical to the pre-move baseline for every fixture", () => {
+    const actual: Record<string, unknown> = {};
+    for (const f of FIXTURES) {
+      actual[f.name] = serializeModel(buildCellModel(f.live, f.input, NOW));
+    }
+    // Byte-identity: serialize both sides with the SAME formatting the baseline
+    // JSON was written with (2-space indent) and compare the resulting strings.
+    const actualStr = JSON.stringify(actual, null, 2);
+    const baselineStr = JSON.stringify(BASELINE, null, 2);
+    expect(actualStr).toBe(baselineStr);
+  });
+
+  // Per-fixture assertions so a drift points at the exact column that moved.
+  for (const f of FIXTURES) {
+    it(`preserves behavior for fixture: ${f.name}`, () => {
+      const actual = serializeModel(buildCellModel(f.live, f.input, NOW));
+      expect(actual).toEqual(BASELINE[f.name]);
+    });
+  }
+});
